requirements.txt
tweepy>=4.14.0
openai>=0.28.0
python-dotenv>=1.0.0
schedule>=1.2.0